+++ src/SimpleCommandBus.php
@@ -3,6 +3,7 @@ declare(strict_types=1);
 
 namespace BSP\CommandBus;
 
+use BSP\CommandBus\Contracts\Command;
 use BSP\CommandBus\Contracts\CommandBus;
 use BSP\CommandBus\Contracts\CommandHandlerMap;
 use BSP\CommandBus\Exception\CommandHandlerIsNotCallable;
@@ -17,7 +18,7 @@ final class SimpleCommandBus implements CommandBus
         $this->handlers = $commandHandlerMap->map();
     }
 
-    public function execute(object $command): void
+    public function execute(Command $command): void
     {
         $commandClass = \get_class($command);